Update 2.0
update list:

Elemental FRU guide is now added

Sidebar:
Sidebar has arrows and left border on sub menu now
elemental button takes you to the guide page
sidebar collapsible works slightly different than pages , as it requires unique ID for each button. Don't ask me why, this is the way

Elemental FRU guide
Created a script to dynamically collapse sub-menu and it can be uncollapsed by clicking on toggle-button attached to it. Now by just adding html this function will work, infinitely
Added a script that will read links and will turn any of those marks into yellow : / -
and will turn https to green

My soul and heart were destroyed trying to make this update
Added a code that lists the titles of the page dynamically, by just adding title="" to anything, it'd be added to the list in order, and you'd be able to click on it to scroll to that portion of the site